Add files via upload
You can play UNO against up to 4 AIs. It's a little difficult to play with no animation of the cards turning over. Also, there might be a few bugs when the deck is reshuffled with the discardPile, but that only ever happened once, and I haven't played/tested enough to get the deck to recombine agian.